+++ Assets/Electricity/Script/Elements/ElecResistance.cs
@@ -2,14 +2,13 @@ using UnityEngine;
 
 public class ElecResistance : ElecDipole
 {
-    public float ohm = 10.0f;
+    public double ohm = 10.0f;
     override public void ComputeIntensity()
     {
         if (IsConnected())
         {
-            float tension = GetVoltage();
+            double tension = GetVoltage();
             intensity = tension / ohm;
-
         }
     }
